% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort.R
\name{aggregate_cohort_on_period_type}
\alias{aggregate_cohort_on_period_type}
\title{Group a cohort by a date period range}
\usage{
aggregate_cohort_on_period_type(cohort, period_type = c("months",
  "periods", "financialyears"), aggregate_by = c("national", "team"),
  period_date_field)
}
\arguments{
\item{cohort}{The cohort being worked upon}

\item{period_type}{A string determining what the gap between
time periods should be. The factor should either be 'months',
'periods' or financialyears'. Months and financialyears are
self-explanatory. Periods represent reporting periods - these were
initially quarterly, then were trimesterly, then quarterly again.
Periods are referred to by their numeric number, rather than a
defined time period to avoid confusion in the calcalations.}

\item{period_date_field}{The column to be used for aggregation.}
}
\value{
The cohort is returned in a \code{tibble}.
}
\description{
This is an internal function used when creating cohorts to group
the data by the chosen period range (ie. months, report periods
or financial years) using the specified column.
}
