% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_audit_output_specs.R
\docType{data}
\name{aggregated_output_ac_72hrs}
\alias{aggregated_output_ac_72hrs}
\title{72hr Audit Compliance Measures}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.}
\usage{
aggregated_output_ac_72hrs
}
\description{
This is a vector of audit measures describing the outputs for
establishing SSNAP audit compliance. Audit compliance is applied
to both 72hr and post-72hr measures.
}
\details{
Internally it is used to create the team72hrKIs and patient72hrKIs
cohorts. It is exposed externally to allow customisation of the
output spec.
}
\keyword{datasets}
