% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_audit_output_specs.R
\docType{data}
\name{aggregated_output_casemix_72hrs}
\alias{aggregated_output_casemix_72hrs}
\title{Casemix 72hr Cohort Aggregation of Measures}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 4 columns.}
\usage{
aggregated_output_casemix_72hrs
}
\description{
This is a vector of audit measures describing the measures used to
look at casemix from the 72hrs Cohorts.
}
\details{
Internally it is used to create the Casemix tab in the portfolio.
}
\keyword{datasets}
