% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_audit_output_specs.R
\docType{data}
\name{aggregated_output_kis_72hrs}
\alias{aggregated_output_kis_72hrs}
\title{Team and Patient 72hr Cohort Key Indicators}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 4 columns.}
\usage{
aggregated_output_kis_72hrs
}
\description{
This is a vector of audit measures describing the outputs for the
team and patient 72hr cohort key indicators.
}
\details{
Internally it is used to create the team72hrKIs and patient72hrKIs
cohorts. It is exposed externally to allow customisation of the
output spec.
}
\keyword{datasets}
