% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_audit_output_specs.R
\docType{data}
\name{aggregated_output_kis_discharge}
\alias{aggregated_output_kis_discharge}
\title{Portfolio Discharge Cohort Key Indicators}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.}
\usage{
aggregated_output_kis_discharge
}
\description{
This is a vector of audit measures describing the measures used to
create the 72hr portfolio. This is a large dataset and so should
only be used for this purpose: it is quicker to create a custom
audit spec for specific purposes.
}
\keyword{datasets}
