% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_audit_output_specs.R
\docType{data}
\name{aggregated_output_kis_team_post72hrs}
\alias{aggregated_output_kis_team_post72hrs}
\title{Team post-72hr (all teams) Cohort Key Indicators}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.}
\usage{
aggregated_output_kis_team_post72hrs
}
\description{
This is a vector of audit measures describing the measures used to
create the all teams post-72hr key indicators. This is used in
performance reporting.
}
\keyword{datasets}
