% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_audit_output_specs.R
\docType{data}
\name{aggregated_output_portfolio_post72hrs_7day}
\alias{aggregated_output_portfolio_post72hrs_7day}
\title{Portfolio post-72hr Cohort 7Day Aggregation of Measures}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.}
\usage{
aggregated_output_portfolio_post72hrs_7day
}
\description{
This is a vector of audit measures describing the measures used to
create the post-72hr portfolio, 7-day cohort. This is a large
dataset and so should only be used for this purpose: it is quicker
to create a custom audit spec for specific purposes.
}
\keyword{datasets}
