% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_audit_output_specs.R
\docType{data}
\name{aggregated_output_portfolio_post72hrs_discharge}
\alias{aggregated_output_portfolio_post72hrs_discharge}
\title{Portfolio post-72hr Cohort Discharge Aggregation of Measures}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.}
\usage{
aggregated_output_portfolio_post72hrs_discharge
}
\description{
This is a vector of audit measures describing the measures used to
create the post-72hr portfolio, discharge cohort. This is a large
dataset and so should only be used for this purpose: it is quicker
to create a custom audit spec for specific purposes.
}
\keyword{datasets}
