% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamcodetoname.R
\name{append_team_data_from_teamcode}
\alias{append_team_data_from_teamcode}
\title{append_team_data_from_teamcode}
\usage{
append_team_data_from_teamcode(data_table)
}
\arguments{
\item{data_table}{A tibble (optimised data frame) containing S1PostcodeOut.}
}
\value{
Data table updated with Team names in TeamName, TransferToTeamName, TransferFromTeamName,
and S7TransferHospitalName
}
\description{
\code{teamcodetoname} is a macro (rather than a function) which converts team codes in the raw SSNAP
data to team names.
}
\details{
This macro takes as input a tibble (an optimised data frame; see the tibble package).

The function API needs nailing down more firmly - at present it knows of three hard-coded column
names (TeamName, TransferFrom and TransferTo) but it should be able to work more generically in
future if we provide it with a column name to translate and an output name.

It also requires Unit tests which currently we don't have. Take this function as a 'work in progress'
It will not win an R beauty contest any time soon.
}
\author{
Andrew Hill, \email{andrew.hill@doctors.org.uk}
}
