% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_output_spec.R
\name{as_audit_output_spec}
\alias{as_audit_output_spec}
\title{Coerces a tibble to become an audit output spec.}
\usage{
as_audit_output_spec(x)
}
\arguments{
\item{x}{The audit output specification written as a tibble.}
}
\value{
A validated audit_output_spec object.
}
\description{
This function is used to create an audit output specification
from a tibble. It will check that we have two columns named name
and type. Both must contain only strings. The type column must
only contain type keywords. If successful, the tibble is returned
as an audit_output_spec object.
}
