% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_measure.R
\name{audit_measure}
\alias{audit_measure}
\alias{audit_measure_continuous}
\alias{audit_measure_discrete}
\title{Creates a new audit_measure object}
\usage{
audit_measure(stem_name, description, exclusions = NULL, numerators,
  numerator_descriptors = NULL, is_key_indicator = FALSE,
  reference_id = NULL, csv_columns = NULL,
  measure_type = c("discrete", "continuous"))

audit_measure_continuous(stem_name, description, exclusions = NULL,
  numerators, numerator_descriptors = NULL, csv_columns = NULL)

audit_measure_discrete(stem_name, description, exclusions = NULL,
  numerators, numerator_descriptors = NULL, csv_columns = NULL)
}
\arguments{
\item{stem_name}{The name of the new audit measure}

\item{description}{A short textual description of what the audit
measure is intended to measure.}

\item{exclusions}{Any exclusions from the population expressed as
an expression.}

\item{numerators}{One or more numerators for the measure. This is
either the variable if we are looking at a continuous measure, or
the pass / fail criteria expressed as a boolean expression.
Results can be broken down by using a series of numerators.}

\item{numerator_descriptors}{A descriptor for each numerator. These
are optional but if you provide one you must provide one for every
numerator.}

\item{is_key_indicator}{Is this measure a key indicator?}

\item{reference_id}{A unique reference number for this audit
measure. Different audits use different reference schemes. We
don't use this reference for calculations, but it is included in
metadata and may be used for some of the outputs.}

\item{csv_columns}{a character vector of all the columns from the
raw CSV file needed to make this audit measure.}

\item{measure_type}{Describes the type of variable - whether it is
discrete (usually TRUE/FALSE), or continuous (where medians and
quartiles may be used to describe the output).}
}
\value{
The created audit_measure object.
}
\description{
This family of functions are used to create a new audit_measure
object.
Audit measure objects have two subclasses: audit_measure_continuous
and audit_measure_discrete. You can create both using the
audit_measure function (to allow us to create multiple objects at
once, for example from a script).
}
\details{
Audit measure objects describe the rules which will be applied to
a cohort to derive a set of aggregated output measures.
The denominator for these measures will be the cohort that they
are applied over.

Any exclusions from the cohort are specified in exclusions.
Each measure can have one or more numerators. This allows measures
to be described in groups or a breakdown of results into
categories.
}
