% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_bank_holidays.R
\docType{data}
\name{bank_holidays_engwales}
\alias{bank_holidays_engwales}
\title{A vector of the bank holidays in England and Wales}
\format{An object of class \code{Date} of length 65.}
\usage{
bank_holidays_engwales
}
\description{
A list of bank holidays in England and Wales stored as a vector
of Dates, taken from https://www.gov.uk/bank-holidays. These are
needed to analyse differences between normal working hours,
weekends and public holidays. They are accurate up to the end of
2019; they will need periodical maintenance to add the dates for
subsequent bank holidays.
}
\keyword{datasets}
