% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_bank_holidays.R
\docType{data}
\name{bank_holidays_ni}
\alias{bank_holidays_ni}
\title{A vector of the bank holidays in Northern Ireland}
\format{An object of class \code{Date} of length 81.}
\usage{
bank_holidays_ni
}
\description{
A list of bank holidays in Northern Ireland stored as a vector
of Dates, taken from https://www.gov.uk/bank-holidays. These are
needed to analyse differences between normal working hours,
weekends and public holidays. They are accurate up to the end of
2019; they will need periodical maintenance to add the dates for
subsequent bank holidays.
}
\details{
NB these are England and Wales bank holidays plus St Patrick's Day
and Battle of the Boyne.
}
\keyword{datasets}
