% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_data_definition.R
\name{cohort_data_definition}
\alias{cohort_data_definition}
\title{Creates a new cohort_data_definition object}
\usage{
cohort_data_definition(cohort_definition, audit_output_spec)
}
\arguments{
\item{cohort_definition}{A cohort definition object, built using
the \code{cohort_definition} function.}

\item{audit_output_spec}{An output definition, built using the
\code{create_output} function.}
}
\value{
The created cohort_definition object.
}
\description{
This function is used to create a new cohort_data_definition
object.
Results from our audit take two forms - either we want to collect
aggregated data from a temporally grouped cohort of patients; or
we want access to a subset of columns from that cohort.
Both the cohorts and the data that we want tend to have a
many-to-many relationship with each other - ie. we often want the
same aggregated results from multiple different cohort types (in
the case of SSNAP team and patient-centred results). In other
cases we want to apply different rules to the same cohorts in
different circumstances (in the case of SSNAP just Key Indicators
or just audit compliance in some cases, whereas on others the
whole portfolio is needed).
Using cohort_data_definitions we can combine all the permutations
we are likely to use commonly; or make our own custom permutations
eg. when creating user filters in a realtime dashboard.
}
