% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort_helpers.R
\name{cohort_slice_first_team}
\alias{cohort_slice_first_team}
\alias{cohort_slice_last_team}
\title{Return first or latest team in a SSNAP patient episode}
\usage{
cohort_slice_first_team(source_data)

cohort_slice_last_team(source_data)
}
\arguments{
\item{source_data}{A tibble of SSNAP data}
}
\value{
the data will be grouped by \code{PatientId}, sliced to
return one record for each patient spell, and then ungrouped.
}
\description{
These functions group the data by \code{PatientId}, and then
slices by either the lowest (first) or highest (last)
\code{ProClinV1Id}, returning either the first or the latest
record of a patient spell.
}
\details{
It is designed for the SSNAP dataset in the sense that it uses
\code{ProClinV1Id} and \code{PatientId} columns, and assumes a
similar record passing model - but could have more generic uses.
}
