% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_measure.R
\name{create_categories}
\alias{create_categories}
\alias{create_categories.audit_measure_continuous}
\alias{create_categories.audit_measure_discrete}
\title{Create categories of an output from an audit_measure}
\usage{
create_categories(x, numerator = NULL, output_type, category = "")

\method{create_categories}{audit_measure_continuous}(x, numerator = NULL,
  output_type, category = "")

\method{create_categories}{audit_measure_discrete}(x, numerator = NULL,
  output_type, category = "")
}
\arguments{
\item{x}{An \code{audit_measure} object.}

\item{numerator}{If the measure has more than one numerator and
only one of the numerators is required, this should be set to the
name of the numerator. If only one numerator is specified in the
measure, or all numerators are to be used, then this should be
omitted (or set to NULL).}

\item{output_type}{Determines which sort of aggregated results
are being requested from the data.}

\item{category}{Name of the category as a string. This will be set
as a blank string if unspecified.}
}
\description{
This function is a companion to create_output and is called by
create_output_tbl. Where create_output creates a set of expressions
for an audit_measure; create_categories outputs a string vector of
the category type for that measure, allowing us to sort the
aggregated data once produced. Unlike the other class methods, an
extra parameter, category, must be supplied: this is the string
category name which will be assigned to all output rows.
}
