% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_measure.R
\name{create_output}
\alias{create_output}
\alias{create_output.audit_measure_continuous}
\alias{create_output.audit_measure_discrete}
\title{Create outputs from an audit_measure object}
\usage{
create_output(x, numerator = NULL, output_type)

\method{create_output}{audit_measure_continuous}(x, numerator = NULL,
  output_type)

\method{create_output}{audit_measure_discrete}(x, numerator = NULL,
  output_type)
}
\arguments{
\item{x}{An \code{audit_measure} object.}

\item{numerator}{If the measure has more than one numerator and
only one of the numerators is required, this should be set to the
name of the numerator. If only one numerator is specified in the
measure, or all numerators are to be used, then this should be
omitted (or set to NULL).}

\item{output_type}{Determines which sort of aggregated results
are being requested from the data:
\itemize{
  \item \strong{median} {Only applicable to
\code{audit_measure_continuous} types. It calculates the median
and returns it in a column named with the \code{audit_measure}'s
stem name followed by Q2, e.g. \strong{myMeasureQ2}.
Multiple numerators are not supported by the median output type
and will generate an error. This may be fixed if there is a need
for it in future. Exclusion criteria are also not currently
handled - this is likely to be addressed in future.}
  \item \strong{quartiles} {This option is similar to \code{median}
but in addition to generating Q2, it also creates the lower and
upper quartiles, suffixed by Q1 and Q3.}
}}
}
\description{
\code{create_output} is a method to produce a set of instructions
to pass to \code{dplyr::summarise}, which is then applied over a
cohort to produce our cohort results. The instructions are passed
as \code{rlang::exprs}, which can be appended to the summary step
using the \code{rlang::!!!} operator.
Output columns have a consistent naming convention.
}
