% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_measure.R
\name{create_output_tbl}
\alias{create_output_tbl}
\title{Create a list of outputs from a table of audit_measure objects}
\usage{
create_output_tbl(outputs_table)
}
\description{
This function operates on a tibble containing the following
columns:
\itemize{
  \item \strong{x} {The output measure as a quosure.}
  \item \strong{numerator} {a numerator name if one is required.}
  \item \strong{output_type} {The type of output required. For
  continuous variables this should be \strong{median} or
  \strong{quartiles}, for discrete ones \strong{pct} or
  \strong{d_n_pct}.}
}
At present the tibble will need to be created in code because we
are referring to the quosure directly; in future though we will
use a list for all the measures and a name to refer to in the list
will be the column in the table.
}
