% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_measure.R
\name{csv_columns_from_output_table}
\alias{csv_columns_from_output_table}
\title{Given an outputs_table containing a list of aggregated data
outputs, produce a list of all the data columns (raw and modified)
needed to create those outputs.}
\usage{
csv_columns_from_output_table(outputs_table)
}
\arguments{
\item{outputs_table}{A table of audit outputs (previously
created using create_output_tbl())}
}
\value{
A character vector of names of the data columns needed
to create the
}
\description{
Given an outputs_table containing a list of aggregated data
outputs, produce a list of all the data columns (raw and modified)
needed to create those outputs.
}
