% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_na_percentage.R
\name{deprecated_na_percentage}
\alias{deprecated_na_percentage}
\title{deprecated_na_percentage}
\usage{
deprecated_na_percentage(field_name)
}
\arguments{
\item{field_name}{The name of the field to calculate the proportion
of non-NA values for.}
}
\description{
\code{deprecated_na_percentage} Calculates the percentage of rows
which are not NA values.
}
\details{
Calculates the percentage of rows which
are not NA values.
}
\author{
Andrew Hill, \email{andrew.hill@doctors.org.uk}
}
