% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_discharge.R
\name{deprecated_time_difference}
\alias{deprecated_time_difference}
\title{internal_calculated_field_functions_discharge}
\usage{
deprecated_time_difference(start_field, end_field, new_field_name, units)
}
\description{
\code{internal_d1ki_calculated_field_functions} is a list of
quosures used to create extra data columns to imported SSNAP data
to support domain 1 (scanning) aggregation calculations.
}
\details{
Calculated fields are returned as a list of functions so that a
single dplyr mutate operation can undertake all the calculated
fields in a single go.

It adds boolean fields for whether the scan was performed within 1
hour and 12 hours, and duplicates the scan time as a Key Indicator
value.
}
\author{
Andrew Hill, \email{andrew.hill@doctors.org.uk}
}
