% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_time_difference.R
\name{internal_time_difference}
\alias{internal_time_difference}
\title{Calculate difference between two date or times inline}
\usage{
internal_time_difference(start_field, end_field, units = "mins")
}
\arguments{
\item{start_field}{The field at the beginning of the time interval.}

\item{end_field}{The field at the end of the time interval.}

\item{units}{This is a string which must match those in the
difftime function (ie. 'minutes', 'hours', 'days' etc)}
}
\description{
Many measures need to work on the difference between two
date/times.
}
\details{
This function calculates the difference between those times, or
returns an NA. The returned value is an integer, which makes
calculations simple.

Always calculate to the lowest precision you want to report - so
if you want to report hours and minutes then ask for the time in
minutes and convert the value later at the printing stage to be
in hours and minutes.

Many measures needing this should ideally have this value stored
in the source data ultimately, so should be recorded for potential
review of database inputs at a later stage.
}
