% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_measure.R
\name{is_audit_measure}
\alias{is_audit_measure}
\alias{is.audit_measure}
\title{Test if the object is an audit_measure}
\usage{
is_audit_measure(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `audit_measure`
class.
}
\description{
This function returns `TRUE` for audit_measures
}
