% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_output_spec.R
\name{is_audit_output_spec}
\alias{is_audit_output_spec}
\title{Test if the object is an audit_output_spec}
\usage{
is_audit_output_spec(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `audit_output_spec`
class.
}
\description{
This function returns `TRUE` for audit_output_specs
}
