% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_definition.R, R/cohort_data_definition.R
\name{is_cohort_definition}
\alias{is_cohort_definition}
\alias{is.cohort_definition}
\title{Test if the object is an cohort_definition}
\usage{
is_cohort_definition(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `cohort_definition`
class.
}
\description{
This function returns `TRUE` for cohort_definitions
}
