% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_kisym.R
\name{kisym}
\alias{kisym}
\title{kisym}
\usage{
kisym(cohort_type, symbol_name)
}
\arguments{
\item{cohort_type}{the prefix - should be 'TC' or 'PC'}

\item{symbol_name}{the name of the symbol that we are going to use,
eg. MySym would produce a symbol name of TCMySym or PCMySym
respectively.}
}
\description{
\code{kisym} is an internally-used function that allows us to
prefix a key indicator symbol name with 'TC' or 'PC' (for
patient-centred or team-centred).
}
\details{
Although it is internal, we have elected not to prefix the
variable name with internal_ because it is used in situations
where having readable code is important, so the name reads
similarly to 'sym' for that reason.
}
\author{
Andrew Hill, \email{andrew.hill@doctors.org.uk}
}
