% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_report_periods.R
\name{report_period_start_date_in_period}
\alias{report_period_start_date_in_period}
\title{report_period_start_date_in_period}
\usage{
report_period_start_date_in_period(period_number)
}
\arguments{
\item{period_number}{The period you want to know the first date of.}
}
\value{
The date, expressed as a Date object.
}
\description{
\code{report_period_start_date_in_period} takes a SSNAP reporting
period number and returns the start date of that period.
}
\details{
SSNAP has gone from a period of quarterly to trimesterly reports
(and back again). So we can't use conventional quarterly reporting
as we have trimesterly reporting for two financial years.

The report_period functions act as conversions between dates and
the reporting periods.
}
\examples{
report_period_start_date_in_period(16)
}
\author{
Andrew Hill, \email{andrew.hill@doctors.org.uk}
}
