% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data_columns.R
\name{source_data_columns}
\alias{source_data_columns}
\title{Find source data columns from a set of expressions}
\usage{
source_data_columns(exprs_list)
}
\arguments{
\item{exprs_list}{A list of expressions that you want to extract
the column names from}
}
\value{
A string vector of column names needed to produce all the
outputs in the expression list.
}
\description{
This function analyses a list of expressions, and aims to extract
all the field names of anything enclosed with
\code{.data[["ColumnName"]]}. When functional, this will be a
vital function for working out which CSV columns we need for any
given set of functions - currently we have to state this by hand
which is error-prone.
}
\details{
The current code needs optimising in order to practically help
improve query speed.
}
