% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_cohort_definitions.R
\docType{data}
\name{ssnap_cohort_definitions}
\alias{ssnap_cohort_definitions}
\title{Define available cohort definitions}
\format{An object of class \code{list} of length 9.}
\usage{
ssnap_cohort_definitions
}
\description{
These functions provide a list of all the standard cohort
definitions. It is possible to create your own cohorts on the fly.
}
\keyword{datasets}
