% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_cohort_filters.R
\docType{data}
\name{ssnap_cohort_filters}
\alias{ssnap_cohort_filters}
\title{Filters used in the creation of cohorts for SSNAP}
\format{An object of class \code{list} of length 9.}
\usage{
ssnap_cohort_filters
}
\description{
These filters tend to be used frequently in cohort generation: it
also helps readability to design them as a list of expressions
which are inserted as the final query becomes much more readable
and filters can be tested individually with records.
}
\keyword{datasets}
