% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_cohorts.R
\docType{data}
\name{ssnap_cohort_names}
\alias{ssnap_cohort_names}
\title{Pre-built SSNAP cohorts}
\format{An object of class \code{character} of length 9.}
\usage{
ssnap_cohort_names
}
\description{
This is a vector of valid cohort names used by the SSNAP dataset.
SSNAP uses pre-defined cohorts which match cohort selection, to
the type of operation to perform and the list of outputs: this
makes using the outputs simpler than using the lower-level
generic audit functions.
}
\details{
\describe{
  \item{Team72HrCohortKIs}{This cohort provides the team centred
  Key Indicators only for records locked to 72h for patients who
  arrived at hospital (or had their stroke in hospital). The
  results are taken from the first record locked to 72hrs for the
  patient. The results are attributed to the first team which
  treated the patient, regardless of which team locked the record
  to 72h (i.e. the second team may have locked the record to 72h,
  but results are attributed to first team).}
  \item{Patient72HrCohortKIs}{This cohort provides the patient
  centred Key Indicators only for the first 72 hours of care and
  is based on records locked to 72h for patients who arrived at
  hospital (or had their stroke in hospital) divided into the
  aggregated time period.}
  \item{Team72HrCohortPortfolio}{This cohort provides the full
  team centred portfolio for records locked to 72h for patients who
  arrived at hospital (or had their stroke in hospital). The
  results are taken from the first record locked to 72hrs for the
  patient. The results are attributed to the first team which
  treated the patient, regardless of which team locked the record
  to 72h (i.e. the second team may have locked the record to 72h,
  but results are attributed to first team).}
  \item{Patient72HrCohortPortfolio}{This cohort provides the
  patient centred portfolio for the first 72 hours of care and
  is based on records locked to 72h for patients who arrived at
  hospital (or had their stroke in hospital) divided into the
  aggregated time period.}
  \item{TeamDischargeCohortKIs}{Results attributed to discharging
  team are attributed to the team which discharged the patient
  from inpatient care. The team centred post-72h results are based
  on records locked to discharge for patients who were discharged
  from inpatient care between the aggregated time periods.}
  \item{Team7DayCohortKIs}{NOT YET COMPLETE For the team centred
  post-72h results, measures are attributed to teams depending on
  the point at which they treated the patient along the inpatient
  pathway. Results attributed to the ‘7 day team’ are attributed
  to the team which had the patient in their care at 7 days
  following clock start (or, if the length of stay as an inpatient
  was less than 7 days, the team which discharged the patient from
  inpatient care). These results include measures which are
  considered to be most appropriately designated to the team which
  had the patient in their care at 7 days, but does not
  necessarily indicate that the care was received within 7 days.
  For instance, one measure in this section is whether the patient
  had a urinary continence plan by discharge. It is attributed to
  the team which had the patient at 7 days, but the measure is
  whether the patient had the plan by discharge.}
  \item{TeamPost72HrsAllTeamsCohortKIs}{Results attributed to all
  teams are for measures which are answered for every patient by
  every team along the pathway (therapy intensity, rehab goal
  setting, length of stay in hospital and on stroke unit and the
  discharge/transfer destination). These results are based only on
  what the team provided rather than what the patient received
  across the whole pathway, e.g. the team centred length of stay
  is the length of stay at each particular team compared to the
  patient centred length of stay which is the length of stay the
  patient had across all teams. The team centred post-72h all
  teams results are based on records locked to discharge or where
  a transfer to another team has been actioned between each
  aggregating period. This includes all records which have either
  been discharged out of inpatient care or transferred to another
  inpatient team.}
  \item{PatientPost72HrsCohortKIs}{NOT YET COMPLETE This section
  shows the patient centred results for the care between 72 hours
  and discharge from inpatient care. It is attributed to all teams
  which treated the patient at any point in their care. This means
  that a team which only treated the patient during the first 72h
  will still have the results for this patient's care between 72
  hours and discharge from inpatient care. We hope that this will
  encourage an open dialogue between teams treating patients along
  a care pathway and that teams treating the patient initially
  reflect on the continuing care they receive, as this will also
  impact upon the initial team's longer term outcome results. The
  patient centred post-72h results are based on records locked to
  discharge for patients who were discharged between each
  aggregating period.}
 }
}
\keyword{datasets}
