% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_fake_data-data.R
\docType{data}
\name{ssnap_fake_data}
\alias{ssnap_fake_data}
\title{SSNAP fake data (for testing / development)}
\format{An object of class \code{tibble}; the data is created
from a SSNAP CSV export of the fake hospital data and run through
\code{ssnapinterface::read_audit_csv()} to generate the dataset.}
\source{
The \href{http://www.strokeaudit.org}{Stroke Sentinel 
National Audit Programme} have created these records. It is
possible to contribute to this dataset by contacting the team
directly.
}
\usage{
data(ssnap_fake_data)
}
\description{
Data that has been entered into SSNAP under the fictitious
'Fake Hospital'. Please NB this is not meant to be a
representative dataset: it is mainly used to test functions and
has examples to test any problems.
}
\examples{
data(ssnap_fake_data)
}
\references{
SSNAP Website
(\href{http://www.strokeaudit.org}{Stroke Sentinel National Audit 
Programme})
}
\keyword{datasets}
