% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_measures.R
\docType{data}
\name{ssnap_measures}
\alias{ssnap_measures}
\title{Define available auditing measures}
\format{An object of class \code{list} of length 191.}
\usage{
ssnap_measures
}
\description{
These functions provide a list of all the available auditing
measures. The master list is retrieved using \code{ssnap_measures}
which is an amalgamation of all lists; lists of measures intended
for a particular cohort group are stored using the other functions
for code legibility and/or ease of access if you know you only
need a subset of the measures.
}
\keyword{datasets}
