% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_report_periods.R
\name{ssnap_report_period}
\alias{ssnap_report_period}
\title{Convert a date to the report number active at that time.}
\usage{
ssnap_report_period(date)
}
\arguments{
\item{date}{The date you want to know the reporting period for as
a Date object.}
}
\value{
The reporting period, expressed as an integer value.
}
\description{
SSNAP has gone from a period of quarterly to trimesterly reports
(and back again). So we can't use conventional quarterly reporting
as we have trimesterly reporting for two financial years.
This function calculates the current report number for a given
date.
}
\examples{
ssnap_report_period(as.Date('2014-02-01'))
}
