% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_report_periods.R
\docType{data}
\name{ssnap_report_period_series}
\alias{ssnap_report_period_series}
\title{Tibble Time time series for SSNAP report periods}
\format{An object of class \code{Date} of length 70.}
\usage{
ssnap_report_period_series
}
\description{
SSNAP has gone from a period of quarterly to trimesterly reports
(and back again). So we can't use conventional quarterly reporting
as we have trimesterly reporting for two financial years.
This sets up a series of time series using
\code{tibbletime::create_series}. Please note that the series
currently ends in 2030 - the end date will need updating if the
code is to be used beyond that date.
}
\examples{
ssnap_report_period_series
}
\keyword{datasets}
