% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_report_periods.R
\docType{data}
\name{ssnap_report_year_series}
\alias{ssnap_report_year_series}
\title{Tibble Time time series for SSNAP report years}
\format{An object of class \code{Date} of length 19.}
\usage{
ssnap_report_year_series
}
\description{
SSNAP follows the UK financial year calendar of April 1st year
start. \code{ssnap_report_year_series} returns a time series using
\code{tibbletime::create_series} following the SSNAP yearly
reports. Please note that the series currently ends in 2030 - the
end date will need updating if the code is to be used beyond that
date.
}
\examples{
ssnap_report_year_series
}
\keyword{datasets}
