% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_values.R
\name{summary_values_ssnap_scores}
\alias{summary_values_ssnap_scores}
\title{Summary Values For SSNAP Scores}
\usage{
summary_values_ssnap_scores(cohort_type)
}
\description{
These calculations are used to produce scores for individual key
indicators and overall domains on SSNAP.
}
