% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_values.R
\docType{data}
\name{summary_values_therapy_compliance}
\alias{summary_values_therapy_compliance}
\title{Summary Values For Therapy Compliance}
\format{An object of class \code{list} of length 3.}
\usage{
summary_values_therapy_compliance
}
\description{
The team post-72hr cohort (Key Indicators in Domains 5, 6 and 7)
use a set of derived summary statistics to calculate the
percentage of compliance with therapy.
}
\keyword{datasets}
